updated hero list & fixed folder structure
