+++ angular-tuorial/angular.json
@@ -34,9 +34,9 @@
 						],
 						"stylePreprocessorOptions": {
 							"includePaths": [
-							  "src/styles"
+								"src/styles"
 							]
-						  },
+						},
 						"scripts": []
 					},
 					"configurations": {
@@ -74,6 +74,7 @@
 				},
 				"serve": {
 					"builder": "@angular-devkit/build-angular:dev-server",
+					
 					"configurations": {
 						"production": {
 							"browserTarget": "angular-tuorial:build:production"
