--- angular-tuorial/src/app/app.module.ts
@@ -1,51 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-
-import { AppComponent } from './app.component';
-import { HeroesComponent } from './heroes/heroes.component';
-import { APP_BASE_HREF } from '@angular/common';
-
-import { FormsModule } from '@angular/forms';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
-import { MesagesComponent } from './mesages/mesages.component';
-import { AppRoutingModule } from './app-routing.module';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { HttpClientModule } from '@angular/common/http';
-import { HttpClientInMemoryWebApiModule} from 'angular-in-memory-web-api';
-import { InMemoryDataService } from './in-memory-data.service';
-import { HeroSearchComponent } from './hero-search/hero-search.component';
-import { HeaderComponent } from './header/header.component';
-import { MainComponent } from './main/main.component';
-import { SidebarComponent } from './sidebar/sidebar.component';
-import { NewsComponent } from './news/news.component';
-
-
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeroesComponent,
-    HeroDetailComponent,
-    MesagesComponent,
-    DashboardComponent,
-    HeroSearchComponent,
-    HeaderComponent,
-    MainComponent,
-    SidebarComponent,
-    NewsComponent
-  ],
-  imports: [
-    BrowserModule,
-    FormsModule,
-    AppRoutingModule,
-    HttpClientModule,
-    HttpClientInMemoryWebApiModule.forRoot(
-      InMemoryDataService, { dataEncapsulation: false }
-    )
-  ],
-  providers: [ ],
-  bootstrap: [AppComponent]
-
- 
-})
-export class AppModule { }
-
