+++ angular-tuorial/src/app/components/header/hero-search/hero-search.component.scss
