+++ angular-tuorial/src/app/components/header/hero-search/hero-search.component.ts
@@ -6,8 +6,8 @@ import {
    debounceTime, distinctUntilChanged, switchMap
  } from 'rxjs/operators';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../../../interfaces/hero';
+import { HeroService } from '../../../services/hero.service';
 
 @Component({
   selector: 'app-hero-search',
