+++ angular-tuorial/src/app/components/main/heroes/hero-detail/hero-detail.component.html
@@ -3,9 +3,9 @@
     <div><span>id: </span>{{hero.id}}</div>
     <div>
         <label for="hero-name">Hero name: </label>
-        <input id="hero-name" [(ngModel)]="hero.name" placeholder="name">
+        <input #heroName id="hero-name" [(ngModel)]="hero.name" placeholder="name">
     </div>
 </div>
 
 <!-- <button (click)="goBack()">go back</button> -->
-<button (click)="save()">Save</button>
+<button *ngIf="hero" (click)="save()">Save</button>
