+++ angular-tuorial/src/app/components/main/heroes/hero-detail/hero-detail.component.scss
