+++ angular-tuorial/src/app/components/main/heroes/heroes.component.scss
