+++ angular-tuorial/src/app/components/main/main.component.ts
