+++ angular-tuorial/src/app/components/main/mesages/mesages.component.ts
@@ -1,4 +1,4 @@
-import { MessageService } from './../message.service';
+import { MessageService } from '../../../services/message.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
