+++ angular-tuorial/src/app/components/sidebar/dashboard/dashboard.component.ts
@@ -1,8 +1,8 @@
 import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
-import { MessageService } from '../message.service';
+import { Hero } from '../../../interfaces/hero';
+import { HeroService } from '../../../services/hero.service';
+import { MessageService } from '../../../services/message.service';
 
 @Component({
   selector: 'app-dashboard',
