--- angular-tuorial/src/app/hero-detail/hero-detail.component.ts
@@ -1,36 +1,0 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Location } from '@angular/common';
-
-import { HeroService } from '../hero.service';
-import { Hero } from '../hero';
-
-@Component({
-  selector: 'app-hero-detail',
-  templateUrl: './hero-detail.component.html',
-  styleUrls: ['./hero-detail.component.scss']
-})
-export class HeroDetailComponent implements OnInit {
-  hero: Hero = {name: '', id: 0};
-  constructor(
-    private route: ActivatedRoute,
-    private heroService: HeroService,
-    private location: Location
-  ) {}
-
-  ngOnInit(): void {
-    this.route.params.subscribe(params => {this.getHero(params['id']);});
-  }
-
-  getHero(id: number): void {
-    this.heroService.getHero(id)
-     .subscribe(hero => this.hero = hero);
-  }
-
-  save(): void {
-    if (this.hero) {
-      this.heroService.updateHero(this.hero);
-        //.subscribe(() => this.location.back());
-    }
-  }
-}
