+++ angular-tuorial/src/app/interfaces/news-rss.ts
