+++ angular-tuorial/src/app/modules/app-routing.module.ts
@@ -1,10 +1,10 @@
 
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HeroesComponent } from './heroes/heroes.component';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
-import { MesagesComponent } from './mesages/mesages.component';
-import { NewsComponent } from './news/news.component';
+import { HeroesComponent } from '../components/main/heroes/heroes.component';
+import { HeroDetailComponent } from '../components/main/heroes/hero-detail/hero-detail.component';
+import { MesagesComponent } from '../components/main/mesages/mesages.component';
+import { NewsComponent } from '../components/main/news/news.component';
 
 const routes: Routes = [
   { path: 'heroes', component: HeroesComponent, children: [
