+++ angular-tuorial/src/app/services/in-memory-data.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { InMemoryDbService } from 'angular-in-memory-web-api';
-import { Hero } from './hero';
+import { Hero } from '../interfaces/hero';
 
 @Injectable({
   providedIn: 'root',
